fix: share follow result types and enable Node typings
